Add files via upload
This script provides information about the process for geo-referencing the units of analysis in the model. This process is important for the calculation of potential elements of interest such as the distances or the closer neighbours to the unit of analysis in the model. Also geo-referenced units are important for geographic data visualisations.